requirements.txt
scrapy
begins